task4

+++ 0x1A-hash_tables/4-hash_table_get.c
@@ -7,7 +7,7 @@
  */
 char *hash_table_get(const hash_table_t *ht, const char *key)
 {
-	unsigned int index;
+	unsigned long int index;
 	hash_node_t *ptr;
 
 	if (ht == NULL || key == NULL || ht->array == NULL ||
